# swagger/portal-swagger.yaml
swagger: '2.0'
info:
  version: '0.9'
  title: Portal
  description: >-
    The API of the Portal (the UI). All of these API calls may return different results
    depending on the current session; you will need to login to get access to restricted
    resources, e.g. to APIs or Content. Most of these end points require an
    `accept: application/json` header to return machine readable results, otherwise you
    will be served the corresponding HTML page. The data you get back from these end points
    corresponds 1:1 with the data visible on the pages; in most cases it's just that the
    rendering has been left out.
...
paths:
...
  /apis/{apiId}:
    get:
      description: >-
        Retrieve information on a specific API, including applications and subscriptions
      tags:
        - api
      parameters:
        - name: apiId
          required: true
          in: path
          description: The API ID to retrieve information on
          type: string
      responses:
        200:
          description: Successful response
...
  /applications/{appId}:
    get:
      description: Retrieve information on an application; includes API credentials
      tags:
        - application
      parameters:
        - name: appId
          in: path
          required: true
          type: string
          description: The ID of the application to retrieve information on
      responses:
        200:
          description: Successful response
...
  /users/{userId}:
    get:
      description: Retrieve information on a user
      tags:
        - user
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user to retrieve information on
          type: string
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed. You can only retrieve yourself if not admin
    post:
      description: Update a user, or delete password for a user
      tags:
        - user
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user to update
          type: string
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              __action:
                type: string
                description: >-
                  Set to `deletePassword` to delete a user's password, otherwise leave blank
              firstname:
                type: string
              lastname:
                type: string
              password:
                type: string
              __updategroups:
                type: boolean
                description: >-
                  Set to `true` to update the groups. Add the group names as properties in the body, containing the group name as value, e.g. `"dev":"dev"` or `"admin":"admin"`.
      responses:
        200:
          description: Successfully patched user.
        204:
          description: Successfully deleted password.
...